limit maximum retry time when timeout, because the server may timeout infinite times

+++ core/image_downloader.cpp
@@ -13,7 +13,7 @@
 #include <QNetworkProxy>
 
 image_downloader::image_downloader(QObject *parent) :
-    QObject(parent),    
+    QObject(parent),
     downloader_{new qte::net::download_supervisor(this)}
 {
     using namespace qte::net;
@@ -34,7 +34,7 @@ void image_downloader::set_download_request(QStringList big_image_links, QString
     QLOG_INFO()<<__func__<<":big img:"<<big_img_links_.size()<<",small img:"<<small_img_links_.size();
     statistic_.clear();
     img_links_map_.clear();
-    save_at_ = save_at;    
+    save_at_ = save_at;
     statistic_.total_download_ = std::min(static_cast<size_t>(big_img_links_.size()),
                                           max_download);
 }
@@ -102,10 +102,11 @@ void image_downloader::download_finished(image_downloader::download_img_task tas
     if(it != std::end(img_links_map_)){
         auto img_info = it->second;
         img_links_map_.erase(it);
-        if(task->get_is_timeout()){
+        if(task->get_is_timeout() && img_info.timeout_retry_num_ < global_constant::timeout_retry_limit()){
             QLOG_INFO()<<__func__<<":"<<task->get_save_as()<<","<<task->get_url()<<": timeout";
             emit set_statusbar_msg(tr("Waiting reply from the server, please give some patient"));
             remove_file("time out issue:", task);
+            ++img_info.timeout_retry_num_;
             download_image(std::move(img_info));
         }else{
             process_download_image(task, std::move(img_info));
@@ -270,17 +271,20 @@ size_t image_downloader::download_statistic::success() const
 
 image_downloader::img_links_map_value::img_links_map_value() :
     choice_{link_choice::big},
-    retry_num_{0}
+    retry_num_{0},
+    timeout_retry_num_{0}
 {
 
 }
 
 image_downloader::img_links_map_value::img_links_map_value(QString big_img_link, QString small_img_link,
-                                                           image_downloader::link_choice choice, size_t retry_num) :
-    big_img_link_(std::move(big_img_link)),    
-    choice_(choice),
-    retry_num_(retry_num),
-    small_img_link_(std::move(small_img_link))
+                                                           image_downloader::link_choice choice, size_t retry_num,
+                                                           size_t timeout_retry_num) :
+    big_img_link_{std::move(big_img_link)},
+    choice_{choice},
+    retry_num_{retry_num},
+    small_img_link_{std::move(small_img_link)},
+    timeout_retry_num_{timeout_retry_num}
 {
 
 }
